Added shadow mapping shader and example scene state.

+++ include/ZSpaceInput.hpp
@@ -19,7 +19,7 @@ namespace spark
     public:
         ZSpaceInputDevice();
         virtual ~ZSpaceInputDevice();
-        virtual void update( float dt ) override;
+        virtual void update( double dt ) override;
         virtual glm::vec3 getPosition( void ) const override;
         virtual glm::mat4 getTransform( void ) const override;
         virtual bool isButtonPressed( int buttonNumber ) const override;
